+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 5.2.0 / 2020-07-27
 * Implement support for GET /contacts/groups
 * Update lodash import
 * Support GET /resources